[ws-manager] Add Git token to workspace secret
