+++ components/content-service/pkg/initializer/initializer.go
@@ -12,7 +12,6 @@ import (
 	"net/http"
 	"os"
 	"path/filepath"
-	"strconv"
 	"strings"
 	"time"
 
@@ -524,119 +523,4 @@ func PlaceWorkspaceReadyFile(ctx context.Context, wspath string, initsrc csapi.W
 	}
 
 	return nil
-}
-
-// GetCheckoutLocationsFromInitializer returns a list of all checkout locations from the initializer
-func GetCheckoutLocationsFromInitializer(init *csapi.WorkspaceInitializer) []string {
-	var res []string
-	_ = WalkInitializer(nil, init, func(path []string, init *csapi.WorkspaceInitializer) error {
-		switch spec := init.Spec.(type) {
-		case *csapi.WorkspaceInitializer_Git:
-			res = append(res, spec.Git.CheckoutLocation)
-		case *csapi.WorkspaceInitializer_Backup:
-			res = append(res, spec.Backup.CheckoutLocation)
-
-		case *csapi.WorkspaceInitializer_Prebuild:
-			// walkInitializer will visit the Git initializer
-		}
-
-		return nil
-	})
-	return res
-}
-
-const extractedSecretPrefix = "extracted-secret/"
-
-// ExtractSecretsFromInitializer removes secrets to the initializer.
-// This is the counterpart of InjectSecretsToInitializer.
-func ExtractSecretsFromInitializer(init *csapi.WorkspaceInitializer) map[string]string {
-	res := make(map[string]string)
-
-	_ = WalkInitializer([]string{"initializer"}, init, func(path []string, init *csapi.WorkspaceInitializer) error {
-		git, ok := init.Spec.(*csapi.WorkspaceInitializer_Git)
-		if !ok {
-			return nil
-		}
-
-		pwd := git.Git.Config.AuthPassword
-		if pwd == "" || strings.HasPrefix(pwd, extractedSecretPrefix) {
-			return nil
-		}
-
-		name := filepath.Join(path...)
-		res[name] = pwd
-		git.Git.Config.AuthPassword = extractedSecretPrefix + name
-
-		return nil
-	})
-
-	return res
-}
-
-// InjectSecretsToInitializer injects secrets to the initializer. This is the counterpart of ExtractSecretsFromInitializer.
-func InjectSecretsToInitializer(init *csapi.WorkspaceInitializer, secrets map[string]string) error {
-	return WalkInitializer([]string{"initializer"}, init, func(path []string, init *csapi.WorkspaceInitializer) error {
-		git, ok := init.Spec.(*csapi.WorkspaceInitializer_Git)
-		if !ok {
-			return nil
-		}
-
-		pwd := git.Git.Config.AuthPassword
-		if !strings.HasPrefix(pwd, extractedSecretPrefix) {
-			return nil
-		}
-
-		name := strings.TrimPrefix(pwd, extractedSecretPrefix)
-		val, ok := secrets[name]
-		if !ok {
-			return xerrors.Errorf("secret %s not found", name)
-		}
-
-		git.Git.Config.AuthPassword = val
-
-		return nil
-	})
-}
-
-// WalkInitializer walks the initializer structure
-func WalkInitializer(path []string, init *csapi.WorkspaceInitializer, visitor func(path []string, init *csapi.WorkspaceInitializer) error) error {
-	switch spec := init.Spec.(type) {
-	case *csapi.WorkspaceInitializer_Backup:
-		return visitor(append(path, "backup"), init)
-	case *csapi.WorkspaceInitializer_Composite:
-		path = append(path, "composite")
-		err := visitor(path, init)
-		if err != nil {
-			return err
-		}
-		for i, p := range spec.Composite.Initializer {
-			err := WalkInitializer(append(path, strconv.Itoa(i)), p, visitor)
-			if err != nil {
-				return err
-			}
-		}
-		return nil
-	case *csapi.WorkspaceInitializer_Download:
-		return visitor(append(path, "download"), init)
-	case *csapi.WorkspaceInitializer_Empty:
-		return visitor(append(path, "empty"), init)
-	case *csapi.WorkspaceInitializer_Git:
-		return visitor(append(path, "git"), init)
-	case *csapi.WorkspaceInitializer_Prebuild:
-		child := append(path, "prebuild")
-		err := visitor(child, init)
-		if err != nil {
-			return err
-		}
-		for i, g := range spec.Prebuild.Git {
-			err = WalkInitializer(append(child, strconv.Itoa(i)), &csapi.WorkspaceInitializer{Spec: &csapi.WorkspaceInitializer_Git{Git: g}}, visitor)
-			if err != nil {
-				return err
-			}
-		}
-		return nil
-
-	default:
-		return fmt.Errorf("unsupported workspace initializer in walkInitializer - this is a bug in Gitpod")
-	}
 }
